Two fixes:
one avoids printing of null at beginning of an output
second tries to preserve the execution error from Gradle and print it instead of plain "gradle execution failed" message

First of all set the exception and only then mark the status as completed. Otherwise the exception value may not be set when the main thread proceeds.
Start with an empty line, not "null".
Provide as much of information as can be obtained from the exception chain.